Move test clients into eastest pkg
