+++ r_put_store_test.go
@@ -18,7 +18,7 @@ func TestStoreClient_PutStore(t *testing.T) {
 
 	storeName := "random-store"
 
-	err := DefaultServerClient.PutStore(ctx, storeName, &easclient.PutStoreRequest{
+	err := eastest.DefaultServerClient.PutStore(ctx, storeName, &easclient.PutStoreRequest{
 		ConfigurationTemplate: easclient.ConfigurationTemplate{
 			Name: "default",
 			Parameters: []easclient.ConfigurationParameter{
